+++ extensions/copy-path/src/utils/applescript-utils.ts
@@ -88,14 +88,34 @@ tell application "${app}"
     keystroke "c" using command down
     key code 53
   end tell
-  delay 0.1
+  delay 0.2
   set activeTabURL to the clipboard
   return (activeTabURL)
 end tell`;
 
-export const getFirefoxBrowserPath = async (app: string) => {
+export const copyFirefoxBrowserPath = async (app: string) => {
   try {
     return await runAppleScript(scriptFirefoxBrowserPath(app));
+  } catch (e) {
+    return "";
+  }
+};
+
+// safari web app browser
+export const scriptSafariWebAppPath = (app: string) => `
+tell application "${app}" to activate
+tell application "System Events"
+  tell process "${app}"
+    keystroke "c" using {option down, command down}
+  end tell
+  delay 0.2
+  set activeTabURL to the clipboard
+  return (activeTabURL)
+end tell`;
+
+export const copySafariWebAppPath = async (app: string) => {
+  try {
+    return await runAppleScript(scriptSafariWebAppPath(app));
   } catch (e) {
     return "";
   }