+++ extensions/copy-path/CHANGELOG.md
@@ -1,8 +1,8 @@
 # Copy Path Changelog
 
-## [Support Firefox] - {PR_MERGE_DATE}
+## [Support More Browsers] - {PR_MERGE_DATE}
 
-- Support copy url from Firefox browser
+- Support Safari Web App and Firefox browser by a hacker method
 
 ## [Copy Window Title] - 2024-06-06